- first working "embedded" mode - call embedded mode with xsl
git-svn-id: https://server.mycore.de/svn/mycore/trunk@12594 7089d063-bc14-496d-ac90-53de1872eb6c

+++ modules/module-ACL-editor_v2/web/JS/aclEditor.js
@@ -25,7 +25,8 @@ function initPermEditor(){
 	}
 	
 	var delAll = getChildrenById(editor, "input", "delAll")[0];
-	delAll.addEventListener("click", deleteAll, false);
+	if (delAll != null)
+		delAll.addEventListener("click", deleteAll, false);
 	
 }
 
@@ -49,7 +50,8 @@ function initRuleEditor(){
 	}
 	
 	var delAll = getChildrenById(editor, "input", "delAll")[0];
-	delAll.addEventListener("click", deleteAll, false);
+	if (delAll != null)
+		delAll.addEventListener("click", deleteAll, false);
 	
 }
